build: fix the remaining %z format strings and an uninitialised read in idHeap
Follow-up to the earlier warning sweep, which only covered the two files that
happened to appear in the log I was given rather than the whole tree.

MinGW's printf has no 'z' length modifier, so every '%zd'/'%zu' in the engine
printed garbage on Windows and GCC reported each one three times (unknown
conversion character, wrong argument type, too many arguments). Twenty sites
across nine files - RenderWorld, VertexCache, tr_trisurf, AAS_routing,
Event, SysCmds, Anim, Script_Program and Dict - now use %i with an explicit
(int) cast. Every value is a byte count, KB figure or object size that fits an
int comfortably; a side-by-side check confirms the printed text, including
field widths like %6zd -> %6i, is unchanged.

idlib/Heap.cpp: idHeap::Init() called AllocatePage() before assigning
defragBlock, and AllocatePage()'s out-of-memory path reads that pointer and
free()s it. The first page allocation of every heap therefore ran with an
indeterminate defragBlock, which GCC flagged as -Wmaybe-uninitialized through
the idHeap constructor and Mem_Init(). Moved the initialisation above the
first call.

The recvfrom()/sendto() -fpermissive warnings in the same log are already
fixed by 19878185 (the log predates it). A clean build is warning-free and the
core is byte-identical in size.

+++ neo/d3xp/ai/AAS_routing.cpp
@@ -326,9 +326,9 @@ void idAASLocal::RoutingStats( void ) const {
 	gameLocal.Printf( "%6d area cache (%d KB)\n", numAreaCache, totalAreaCacheMemory >> 10 );
 	gameLocal.Printf( "%6d portal cache (%d KB)\n", numPortalCache, totalPortalCacheMemory >> 10 );
 	gameLocal.Printf( "%6d total cache (%d KB)\n", numAreaCache + numPortalCache, totalCacheMemory >> 10 );
-	gameLocal.Printf( "%6d area travel times (%zd KB)\n", numAreaTravelTimes, ( numAreaTravelTimes * sizeof( unsigned short ) ) >> 10 );
-	gameLocal.Printf( "%6d area cache entries (%zd KB)\n", areaCacheIndexSize, ( areaCacheIndexSize * sizeof( idRoutingCache * ) ) >> 10 );
-	gameLocal.Printf( "%6d portal cache entries (%zd KB)\n", portalCacheIndexSize, ( portalCacheIndexSize * sizeof( idRoutingCache * ) ) >> 10 );
+	gameLocal.Printf( "%6d area travel times (%i KB)\n", numAreaTravelTimes, (int)( ( numAreaTravelTimes * sizeof( unsigned short ) ) >> 10 ) );
+	gameLocal.Printf( "%6d area cache entries (%i KB)\n", areaCacheIndexSize, (int)( ( areaCacheIndexSize * sizeof( idRoutingCache * ) ) >> 10 ) );
+	gameLocal.Printf( "%6d portal cache entries (%i KB)\n", portalCacheIndexSize, (int)( ( portalCacheIndexSize * sizeof( idRoutingCache * ) ) >> 10 ) );
 }
 
 /*
