+++ neo/d3xp/script/Script_Program.cpp
@@ -1245,7 +1245,7 @@ byte *idProgram::ReserveMem(int size) {
 	byte *res = &variables[ numVariables ];
 	numVariables += size;
 	if ( numVariables > sizeof( variables ) ) {
-		throw idCompileError( va( "Exceeded global memory size (%zd bytes)", sizeof( variables ) ) );
+		throw idCompileError( va( "Exceeded global memory size (%i bytes)", (int)sizeof( variables ) ) );
 	}
 
 	memset( res, 0, size );
@@ -1811,13 +1811,13 @@ void idProgram::CompileStats( void ) {
 
 	gameLocal.Printf( "Memory usage:\n" );
 	gameLocal.Printf( "     Strings: %d, %d bytes\n", fileList.Num(), stringspace );
-	gameLocal.Printf( "  Statements: %d, %zd bytes\n", statements.Num(), statements.MemoryUsed() );
+	gameLocal.Printf( "  Statements: %d, %i bytes\n", statements.Num(), (int)statements.MemoryUsed() );
 	gameLocal.Printf( "   Functions: %d, %d bytes\n", functions.Num(), funcMem );
 	gameLocal.Printf( "   Variables: %d bytes\n", numVariables );
 	gameLocal.Printf( "    Mem used: %d bytes\n", memused );
-	gameLocal.Printf( " Static data: %zd bytes\n", sizeof( idProgram ) );
+	gameLocal.Printf( " Static data: %i bytes\n", (int)sizeof( idProgram ) );
 	gameLocal.Printf( "   Allocated: %d bytes\n", memallocated );
-	gameLocal.Printf( " Thread size: %zd bytes\n", sizeof( idThread ) );
+	gameLocal.Printf( " Thread size: %i bytes\n", (int)sizeof( idThread ) );
 }
 
 /*
