+++ neo/idlib/Dict.cpp
@@ -648,8 +648,8 @@ idDict::ShowMemoryUsage_f
 ================
 */
 void idDict::ShowMemoryUsage_f( const idCmdArgs &args ) {
-	idLib::common->Printf( "%5zd KB in %d keys\n", globalKeys.Size() >> 10, globalKeys.Num() );
-	idLib::common->Printf( "%5zd KB in %d values\n", globalValues.Size() >> 10, globalValues.Num() );
+	idLib::common->Printf( "%5i KB in %d keys\n", (int)( globalKeys.Size() >> 10 ), globalKeys.Num() );
+	idLib::common->Printf( "%5i KB in %d values\n", (int)( globalValues.Size() >> 10 ), globalValues.Num() );
 }
 
 /*
